dev: rename --skip-all-confirmations to --install
This option is meant to be used only for the first build,
where users follow the installation/migration
instructions and prepare the necessary things like PGroonga
and it's safe to skip previous notes.

Skipping 'Did you stop your server?' in normal
updates should be done with another option,
which will be added accordingly.